% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AF20151130.R
\name{plot.AF}
\alias{plot.AF}
\title{Plot function for objects of class "\code{AF}" from the function \code{\link{AF.ch}}.}
\usage{
\method{plot}{AF}(x, CI = FALSE, confidence.level, CI.transform, xlab, main,
  ...)
}
\arguments{
\item{x}{an object of class \code{AF} from the \code{\link{AF.ch}} function.}

\item{CI}{if TRUE confidence intervals are estimated and ploted in the graph.}

\item{confidence.level}{user-specified confidence level for the confidence intervals. If not specified it defaults to 95 percent. Should be specified in decimals such as 0.95 for 95 percent.}

\item{CI.transform}{user-specified transformation of the Wald confidence interval(s). Options are \code{untransformed}, \code{log} and \code{logit}. If not specified untransformed will be calculated.}

\item{xlab}{label on the x-axis. If not specified the label \emph{"Time"} will be displayed.}

\item{main}{main title of the plot. If not specified the lable \emph{"Estimate of the attributable fraction function"} will be displayed.}

\item{...}{further arguments to be passed to the plot function. See \code{\link[graphics]{plot}}.}
}
\description{
Creates a simple scatterplot for the AF function with time sequence (specified by the user as \code{times} in the \code{\link{AF.ch}} function) on the x-axis and the AF function estimate on the y-axis.
}
\author{
Elisabeth Dahlqwist, Arvid \enc{Sjölander}{Sjolander}
}

