% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFfunctions.R
\name{summary.AF}
\alias{summary.AF}
\title{Summary function for objects of class "\code{AF}".}
\usage{
\method{summary}{AF}(object, digits = max(3L, getOption("digits") - 3L),
  confidence.level, CI.transform, ...)
}
\arguments{
\item{object}{an object of class \code{AF} from \code{\link{AFglm}}, \code{\link{AFcoxph}}, \code{\link{AFclogit}}, \code{\link{AFparfrailty}} or \code{\link{AFivglm}} functions.}

\item{digits}{maximum number of digits.}

\item{confidence.level}{user-specified confidence level for the confidence intervals. If not specified it defaults to 95 percent. Should be specified in decimals such as 0.95 for 95 percent.}

\item{CI.transform}{user-specified transformation of the Wald confidence interval(s). Options are \code{untransformed}, \code{log} and \code{logit}. If not specified untransformed will be calculated.}

\item{...}{further arguments to be passed to the summary function. See \code{\link[base]{summary}}.}
}
\description{
Gives a summary of the AF estimate(s) including z-value, p-value and confidence interval(s).
}
\author{
Elisabeth Dahlqwist, Arvid \enc{Sjölander}{Sjolander}
}
